requirements.txt
# ─── Core deep learning ───────────────────────────────────────────
torch>=2.0.0
torchvision>=0.15.0

# ─── HuggingFace (used for 3-model ensemble in predict.py) ────────
transformers>=4.35.0
accelerate>=0.20.0
huggingface_hub>=0.19.0

# ─── Image processing ─────────────────────────────────────────────
opencv-python>=4.8.0
Pillow>=10.0.0
numpy>=1.24.0

# ─── Web backend ──────────────────────────────────────────────────
fastapi>=0.103.0
uvicorn[standard]>=0.23.0
python-multipart>=0.0.6

# ─── Training-only (not needed to run the API) ────────────────────
timm>=0.9.0
scikit-learn>=1.3.0
matplotlib>=3.7.0
tqdm>=4.66.0

# ─── Video processing (required for video_predict.py) ─────────────
# opencv-python above already handles video decoding via VideoCapture
# No additional package required — OpenCV ships with full codec support
